Etap2 karol parser

See merge request morzelo1/apro2_23l_pro_pt_rozliczenie_wspolnych_wydatkow!6

+++ requests/src/main/java/server/app/Transaction/TransactionController.java
@@ -1,4 +1,5 @@
 package server.app.Transaction;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -6,11 +7,15 @@ import server.app.requests.UserTransactionRequest;
 import server.app.Users.Users;
 import server.app.Users.UsersRepository;
 import server.app.requests.removeTransactionRequest;
-
+import java.math.BigDecimal;
+import java.io.File;
+import java.io.IOException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import static java.util.Objects.isNull;
 
@@ -57,31 +62,69 @@ public class TransactionController {
     public ResponseEntity<String> addTransaction(@RequestBody UserTransactionRequest utRequest){
         Users sender = usersRepository.findByName(utRequest.sender);
         Users receiver = usersRepository.findByName(utRequest.receiver);
-        Transaction transaction = new Transaction(sender,receiver,utRequest.amount, LocalDate.now());
-        transactionrepository.save(transaction);
-        //TransactionGraph trGraph = new TransactionGraph(TransactionGraph.getListOfTransactions(), (ArrayList<Users>) usersRepository.findAll());
-        //transactionrepository.saveAll(TransactionGraph.getListOfTransactions());
-        TransactionGraph.users=((ArrayList<Users>) usersRepository.findAll());
-        //TransactionGraph.listOfTransactions= (ArrayList<Transaction>) transactionrepository.findAll();
-        TransactionGraph.simplify();
-
-        //System.out.println(listOfTransactions2 + "addtransaction");
-        transactionrepository.deleteAll(transactionrepository.findAll());
-        transactionrepository.saveAll(TransactionGraph.simiplifiedList);
+        BigDecimal amount = new BigDecimal(utRequest.amount);
+        if(validateUserInput(utRequest.receiver)){
+            Transaction transaction = new Transaction(sender,receiver,amount, LocalDate.now());
+            transactionrepository.save(transaction);
+            TransactionGraph.users=((ArrayList<Users>) usersRepository.findAll());
+            TransactionGraph.simplify();
+            transactionrepository.deleteAll(transactionrepository.findAll());
+            transactionrepository.saveAll(TransactionGraph.simiplifiedList);
+            ObjectMapper objectMapper = new ObjectMapper();
+            List<Transaction> transactions = new ArrayList<>();
+            transactions = transactionrepository.findAll();
+            try {
+                objectMapper.writeValue(new File("requests/src/main/resources/transactionBackup.json"), transactions);
+            }catch (IOException e){
+                e.printStackTrace();
+            }
         return new ResponseEntity<String>("Record saved successfully", HttpStatus.CREATED);
+        }
+        else return new ResponseEntity<String>("Invalid request", HttpStatus.BAD_REQUEST);
     }
 
     @DeleteMapping("/removeTransaction")
     Transaction removeTransaction(@RequestBody removeTransactionRequest rtRequest){
         Users sender = usersRepository.findByName(rtRequest.sender);
         Users receiver = usersRepository.findByName(rtRequest.receiver);
-        for(Transaction x: transactionrepository.findAll()){
-            if(x.getSender() == sender && x.getReceiver() == receiver){
-                transactionrepository.delete(x);
-                return x;
+        if(validateUserInput(rtRequest.receiver)) {
+            for (Transaction x : transactionrepository.findAll()) {
+                if (x.getSender() == sender && x.getReceiver() == receiver) {
+                    transactionrepository.delete(x);
+                    return x;
+                }
             }
         }
+
         return null;
+    }
+
+    /**
+     * method used to validate the input of transaction amount
+     * accepts only digits and a dot
+     * @param transactionInput
+     * @return false if the transaction input contains an illegal character, true otherwise
+     *//*
+    public boolean validateTransactionInput(String transactionInput){
+        String input = transactionInput;
+        Pattern pattern = Pattern.compile("^[0-9.]");
+        Matcher matcher = pattern.matcher(input);
+        boolean valid = matcher.find();
+        return !valid;
+    }
+    */
+    /**
+     * method used to validate the User's input: username or password
+     * accepts only characters of English alphabet, digits and a dot
+     * @param userInput
+     * @return false if the given input contains an illegal character, true otherwise
+     */
+    public boolean validateUserInput(String userInput){
+        String input = userInput;
+        Pattern pattern = Pattern.compile("^[a-zA-Z0-9.]");
+        Matcher matcher = pattern.matcher(input);
+        boolean valid = matcher.find();
+        return !valid;
     }
 
 }
